Default Submodel Reconfiguration and Options for Simplified PT File

+++ src/fastga_he/models/loops/wing_area_component/wing_area_cl_dep_equilibrium.py
@@ -69,11 +69,20 @@ class UpdateWingAreaLiftDEPEquilibrium(om.ExplicitComponent):
             desc="Boolean to sort the component with proper order for adding subsystem operations",
             allow_none=False,
         )
+        self.options.declare(
+            name="produce_simplified_pt_file",
+            default=False,
+            desc="Boolean to split powertrain architecture into smaller branches",
+            allow_none=False,
+        )
 
     def setup(self):
         if self.options["power_train_file_path"]:
             self.configurator.load(self.options["power_train_file_path"])
-            self.simplified_file_path = self.configurator.produce_simplified_pt_file_copy()
+            if self.options["produce_simplified_pt_file"]:
+                self.simplified_file_path = self.configurator.produce_simplified_pt_file_copy()
+            else:
+                self.simplified_file_path = self.options["power_train_file_path"]
             self.control_parameter_list = self.configurator.get_control_parameter_list()
 
         self.add_input("data:TLAR:v_approach", val=np.nan, units="m/s")
@@ -136,12 +145,15 @@ class UpdateWingAreaLiftDEPEquilibrium(om.ExplicitComponent):
             method="fd",
         )
 
-        if self.options["power_train_file_path"]:
+        if self.options["power_train_file_path"] and self.options["produce_simplified_pt_file"]:
             os.remove(self.simplified_file_path)
 
     def compute(self, inputs, outputs, discrete_inputs=None, discrete_outputs=None):
         if self.options["power_train_file_path"]:
-            self.simplified_file_path = self.configurator.produce_simplified_pt_file_copy()
+            if self.options["produce_simplified_pt_file"]:
+                self.simplified_file_path = self.configurator.produce_simplified_pt_file_copy()
+            else:
+                self.simplified_file_path = self.options["power_train_file_path"]
 
         # First, compute a failsafe value, in case the computation crashes because of the wrong
         # initial guesses of the problem
@@ -175,7 +187,7 @@ class UpdateWingAreaLiftDEPEquilibrium(om.ExplicitComponent):
 
         outputs["wing_area"] = wing_area_approach
 
-        if self.options["power_train_file_path"]:
+        if self.options["power_train_file_path"] and self.options["produce_simplified_pt_file"]:
             # We can now delete the temp .yml we created, just to avoid over-clogging the repo
             os.remove(self.simplified_file_path)
